Rename csdc-orcid to csdc-auth
