+++ cabal.project
@@ -1,4 +1,4 @@
 packages:
   csdc-api
+  csdc-auth
   csdc-base
-  csdc-orcid
